Deboggle: Merge free_energy and free_energy_rabfe modules
