+++ examples/potential_energy.py
@@ -3,10 +3,9 @@ from jax import numpy as np
 from jax import value_and_grad
 
 from timemachine.fe.functional import construct_differentiable_interface
-from timemachine.testsystems.relative import hif2a_ligand_pair
+from timemachine.testsystems.relative import hif2a_ligand_pair_vacuum_edge
 
-ff_params = hif2a_ligand_pair.ff.get_ordered_params()
-unbound_potentials, sys_params, _, coords = hif2a_ligand_pair.prepare_vacuum_edge(ff_params)
+unbound_potentials, sys_params, _, coords = hif2a_ligand_pair_vacuum_edge
 U = construct_differentiable_interface(unbound_potentials)
 box = np.eye(3) * 100
 lam = 0.5
