+++ timemachine/fe/free_energy.py
@@ -1,29 +1,130 @@
-from typing import List
+import math
 
 from jax.config import config
+from rdkit import Chem
 
 config.update("jax_enable_x64", True)
 
+from dataclasses import asdict, dataclass, fields
+from typing import Union
+
 import numpy as np
-from rdkit.Chem import MolToSmiles
 
 from timemachine.fe import topology
 from timemachine.fe.utils import get_romol_conf
 from timemachine.ff.handlers import openmm_deserializer
-from timemachine.lib import LangevinIntegrator
+
+
+@dataclass(eq=False)
+class RABFEResult:
+    mol_name: str
+    dG_complex_conversion: float
+    dG_complex_conversion_error: float
+    dG_complex_decouple: float
+    dG_complex_decouple_error: float
+    dG_solvent_conversion: float
+    dG_solvent_conversion_error: float
+    dG_solvent_decouple: float
+    dG_solvent_decouple_error: float
+
+    def log(self):
+        """print stage summary"""
+        print(
+            "stage summary for mol:",
+            self.mol_name,
+            "dG_complex_conversion (K complex)",
+            self.dG_complex_conversion,
+            "dG_complex_conversion_err",
+            self.dG_complex_conversion_error,
+            "dG_complex_decouple (E0 + A0 + A1 + E1)",
+            self.dG_complex_decouple,
+            "dG_complex_decouple_err",
+            self.dG_complex_decouple_error,
+            "dG_solvent_conversion (K solvent)",
+            self.dG_solvent_conversion,
+            "dG_solvent_conversion_err",
+            self.dG_solvent_conversion_error,
+            "dG_solvent_decouple (D)",
+            self.dG_solvent_decouple,
+            "dG_solvent_decouple_err",
+            self.dG_solvent_decouple_error,
+        )
+
+    def __eq__(self, other: "RABFEResult") -> bool:
+        if not isinstance(other, RABFEResult):
+            return NotImplemented
+        equal = True
+        for field in fields(self):
+            self_val = getattr(self, field.name)
+            other_val = getattr(other, field.name)
+            # Python doesn't consider nan == nan to be true
+            if field.type is float and math.isnan(self_val) and math.isnan(other_val):
+                continue
+            equal &= self_val == other_val
+        return equal
+
+    @classmethod
+    def _convert_field_to_sdf_field(cls, field_name: str) -> str:
+        if field_name == "mol_name":
+            cleaned_name = "_Name"
+        else:
+            cleaned_name = field_name.replace("_", " ")
+        return cleaned_name
+
+    @classmethod
+    def from_mol(cls, mol: Chem.Mol):
+        field_names = fields(cls)
+
+        kwargs = {}
+        for field in field_names:
+            field_name = cls._convert_field_to_sdf_field(field.name)
+            val = mol.GetProp(field_name)
+            val = field.type(val)
+            kwargs[field.name] = val
+        return RABFEResult(**kwargs)
+
+    def apply_to_mol(self, mol: Chem.Mol):
+        results_dict = asdict(self)
+        results_dict.update(
+            {
+                "dG_bind": self.dG_bind,
+                "dG_bind_err": self.dG_bind_err,
+            }
+        )
+        for field, val in results_dict.items():
+            field_name = self._convert_field_to_sdf_field(field)
+            mol.SetProp(field_name, str(val))
+
+    @property
+    def dG_complex(self):
+        """effective free energy of removing from complex"""
+        return self.dG_complex_conversion + self.dG_complex_decouple
+
+    @property
+    def dG_solvent(self):
+        """effective free energy of removing from solvent"""
+        return self.dG_solvent_conversion + self.dG_solvent_decouple
+
+    @property
+    def dG_bind(self):
+        """the final value we seek is the free energy of moving
+        from the solvent into the complex"""
+        return self.dG_solvent - self.dG_complex
+
+    @property
+    def dG_bind_err(self):
+        errors = np.asarray(
+            [
+                self.dG_complex_conversion_error,
+                self.dG_complex_decouple_error,
+                self.dG_solvent_conversion_error,
+                self.dG_solvent_decouple_error,
+            ]
+        )
+        return np.sqrt(np.sum(errors ** 2))
 
 
 class BaseFreeEnergy:
-    @staticmethod
-    def _get_integrator(combined_masses):
-        """
-        Get a integrator. The resulting impl must be bound to a python handle
-        whose lifetime is concurrent with that of the context.
-        """
-        seed = np.random.randint(np.iinfo(np.int32).max)
-
-        return LangevinIntegrator(300.0, 1.5e-3, 1.0, combined_masses, seed)
-
     @staticmethod
     def _get_system_params_and_potentials(ff_params, topology):
 
@@ -44,33 +145,10 @@ class BaseFreeEnergy:
 
         return final_params, final_potentials
 
-    def prepare_host_edge(self, ff_params, host_system, host_coords):
-        """
-        Prepares the host-edge system
-
-        Parameters
-        ----------
-        ff_params: tuple of np.array
-            Exactly equal to bond_params, angle_params, proper_params, improper_params, charge_params, lj_params
-
-        host_system: openmm.System
-            openmm System object to be deserialized
-
-        host_coords: np.array
-            Nx3 array of atomic coordinates
-
-        Returns
-        -------
-        4 tuple
-            unbound_potentials, system_params, combined_masses, combined_coords
-
-        """
-        raise NotImplementedError("Not implemented")
-
 
 # this class is serializable.
 class AbsoluteFreeEnergy(BaseFreeEnergy):
-    def __init__(self, mol, ff):
+    def __init__(self, mol, top):
         """
         Compute the absolute free energy of a molecule via 4D decoupling.
 
@@ -79,15 +157,14 @@ class AbsoluteFreeEnergy(BaseFreeEnergy):
         mol: rdkit mol
             Ligand to be decoupled
 
-        ff: ff.Forcefield
-            Ligand forcefield
+        top: Topology
+            topology.Topology to use
 
         """
         self.mol = mol
-        self.ff = ff
-        self.top = topology.BaseTopology(mol, ff)
+        self.top = top
 
-    def prepare_host_edge(self, ff_params, host_system, host_coords):
+    def prepare_host_edge(self, ff_params, host_system):
         """
         Prepares the host-edge system
 
@@ -99,18 +176,13 @@ class AbsoluteFreeEnergy(BaseFreeEnergy):
         host_system: openmm.System
             openmm System object to be deserialized
 
-        host_coords: np.array
-            Nx3 array of atomic coordinates
-
         Returns
         -------
-        4 tuple
-            unbound_potentials, system_params, combined_masses, combined_coords
+        3-tuple
+            unbound_potentials, system_params, combined_masses
 
         """
         ligand_masses = [a.GetMass() for a in self.mol.GetAtoms()]
-        ligand_coords = get_romol_conf(self.mol)
-
         host_bps, host_masses = openmm_deserializer.deserialize_system(host_system, cutoff=1.2)
 
         hgt = topology.HostGuestTopology(host_bps, self.top)
@@ -118,16 +190,20 @@ class AbsoluteFreeEnergy(BaseFreeEnergy):
         final_params, final_potentials = self._get_system_params_and_potentials(ff_params, hgt)
 
         combined_masses = np.concatenate([host_masses, ligand_masses])
-        combined_coords = np.concatenate([host_coords, ligand_coords])
 
-        return final_potentials, final_params, combined_masses, combined_coords
+        return final_potentials, final_params, combined_masses
+
+    def prepare_combined_coords(self, host_coords):
+        # TODO ...
+        ligand_coords = get_romol_conf(self.mol)
+        return np.concatenate([host_coords, ligand_coords])
 
 
 # this class is serializable.
 class RelativeFreeEnergy(BaseFreeEnergy):
-    def __init__(self, single_topology: topology.SingleTopology, label=None, complex_path=None):
-        self.top = single_topology
-        self.label = label
+    def __init__(self, top: Union[topology.SingleTopology, topology.DualTopology], label=None, complex_path=None):
+        self.top = top
+        self.label = label  # TODO: Do we need these?
         self.complex_path = complex_path
 
     @property
@@ -139,48 +215,10 @@ class RelativeFreeEnergy(BaseFreeEnergy):
         return self.top.mol_b
 
     @property
-    def core(self):
-        return self.top.core
-
-    @property
     def ff(self):
         return self.top.ff
 
-    def _get_integrator(self, combined_masses):
-        """
-        Get a integrator. The resulting impl must be bound to a python handle
-        whose lifetime is concurrent with that of the context.
-        """
-        seed = np.random.randint(np.iinfo(np.int32).max)
-
-        return LangevinIntegrator(300.0, 1.5e-3, 1.0, combined_masses, seed)
-
-    def prepare_vacuum_edge(self, ff_params):
-        """
-        Prepares the vacuum system.
-        Parameters
-        ----------
-        ff_params: tuple of np.array
-            Exactly equal to bond_params, angle_params, proper_params, improper_params, charge_params, lj_params
-        Returns
-        -------
-        4 tuple
-            unbound_potentials, system_parameters, combined_masses, combined_coords
-        """
-        ligand_masses_a = [a.GetMass() for a in self.mol_a.GetAtoms()]
-        ligand_masses_b = [b.GetMass() for b in self.mol_b.GetAtoms()]
-
-        ligand_coords_a = get_romol_conf(self.mol_a)
-        ligand_coords_b = get_romol_conf(self.mol_b)
-
-        final_params, final_potentials = self._get_system_params_and_potentials(ff_params, self.top)
-
-        combined_masses = np.mean(self.top.interpolate_params(ligand_masses_a, ligand_masses_b), axis=0)
-        combined_coords = np.mean(self.top.interpolate_params(ligand_coords_a, ligand_coords_b), axis=0)
-
-        return final_potentials, final_params, combined_masses, combined_coords
-
-    def prepare_host_edge(self, ff_params, host_system, host_coords):
+    def prepare_host_edge(self, ff_params, host_system):
         """
         Prepares the host-edge system
         Parameters
@@ -189,91 +227,41 @@ class RelativeFreeEnergy(BaseFreeEnergy):
             Exactly equal to bond_params, angle_params, proper_params, improper_params, charge_params, lj_params
         host_system: openmm.System
             openmm System object to be deserialized
-        host_coords: np.array
-            Nx3 array of atomic coordinates
+
         Returns
         -------
-        4 tuple
-            unbound_potentials, system_params, combined_masses, combined_coords
+        3-tuple
+            unbound_potentials, system_params, combined_masses
+
         """
 
         ligand_masses_a = [a.GetMass() for a in self.mol_a.GetAtoms()]
         ligand_masses_b = [b.GetMass() for b in self.mol_b.GetAtoms()]
-
-        # extract the 0th conformer
-        ligand_coords_a = get_romol_conf(self.mol_a)
-        ligand_coords_b = get_romol_conf(self.mol_b)
 
         host_bps, host_masses = openmm_deserializer.deserialize_system(host_system, cutoff=1.2)
 
         hgt = topology.HostGuestTopology(host_bps, self.top)
 
         final_params, final_potentials = self._get_system_params_and_potentials(ff_params, hgt)
+
         if isinstance(self.top, topology.SingleTopology):
             combined_masses = np.concatenate(
                 [host_masses, np.mean(self.top.interpolate_params(ligand_masses_a, ligand_masses_b), axis=0)]
             )
+        else:
+            combined_masses = np.concatenate([host_masses, ligand_masses_a, ligand_masses_b])
+
+        return final_potentials, final_params, combined_masses
+
+    def prepare_combined_coords(self, host_coords):
+        # TODO ...
+        ligand_coords_a = get_romol_conf(self.mol_a)
+        ligand_coords_b = get_romol_conf(self.mol_b)
+
+        if isinstance(self.top, topology.SingleTopology):
             combined_coords = np.concatenate(
                 [host_coords, np.mean(self.top.interpolate_params(ligand_coords_a, ligand_coords_b), axis=0)]
             )
         else:
-            combined_masses = np.concatenate([host_masses, ligand_masses_a, ligand_masses_b])
             combined_coords = np.concatenate([host_coords, ligand_coords_a, ligand_coords_b])
-
-        return final_potentials, final_params, combined_masses, combined_coords
-
-
-class RBFETransformIndex:
-    """Builds an index of relative free energy transformations to use
-    with construct_mle_layer
-    """
-
-    def __len__(self):
-        return len(self._indices)
-
-    def __init__(self):
-        self._indices = {}
-
-    def build(self, refs: List[RelativeFreeEnergy]):
-        for ref in refs:
-            self.get_transform_indices(ref)
-
-    def get_transform_indices(self, ref: RelativeFreeEnergy) -> List[int]:
-        return self.get_mol_idx(ref.mol_a), self.get_mol_idx(ref.mol_b)
-
-    def get_mol_idx(self, mol):
-        hashed = self._mol_hash(mol)
-        if hashed not in self._indices:
-            self._indices[hashed] = len(self._indices)
-        return self._indices[hashed]
-
-    def _mol_hash(self, mol):
-        return MolToSmiles(mol)
-
-
-def construct_lambda_schedule(num_windows):
-    """Generate a length-num_windows list of lambda values from 0.0 up to 1.0
-
-    Notes
-    -----
-    manually optimized by YTZ
-    """
-
-    A = int(0.35 * num_windows)
-    B = int(0.30 * num_windows)
-    C = num_windows - A - B
-
-    # Empirically, we see the largest variance in std <du/dl> near the endpoints in the nonbonded
-    # terms. Bonded terms are roughly linear. So we add more lambda windows at the endpoint to
-    # help improve convergence.
-    lambda_schedule = np.concatenate(
-        [
-            np.linspace(0.0, 0.25, A, endpoint=False),
-            np.linspace(0.25, 0.75, B, endpoint=False),
-            np.linspace(0.75, 1.0, C, endpoint=True),
-        ]
-    )
-
-    assert len(lambda_schedule) == num_windows
-
-    return lambda_schedule
+        return combined_coords
